+++ LD_GG/src/main/java/com/ld/gg/controller/summoner/SummonerController.java
@@ -1,4 +1,4 @@
-package com.ld.gg.controller;
+package com.ld.gg.controller.summoner;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
 
-import com.ld.gg.dto.SummonerDto;
+import com.ld.gg.dto.summoner.SummonerDto;
 import com.ld.gg.service.SummonerService;
 
 @Controller
