+++ LD_GG/src/main/java/com/ld/gg/controller/summoner/SummonerRestController.java
@@ -6,6 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.ld.gg.dto.summoner.RecordDto;
 import com.ld.gg.dto.summoner.SummonerDto;
 import com.ld.gg.service.SummonerService;
